backend/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
filelock==3.14.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fsspec==2024.6.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.31
google-pasta==0.2.0
gpt_2_simple==0.8.1
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
nodeenv==1.6.0
numpy==1.26.4
openai==1.0.0
openpyxl==3.1.3
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pandas-stubs==2.2.2.240603
protobuf==4.25.3
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pymongo==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.28.2
rich==13.7.1
safetensors==0.4.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sympy==1.12.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tf_keras==2.16.0
tokenizers==0.19.1
toposort==1.10
torch==2.3.1
tqdm==4.66.4
transformers==4.41.2
types-pytz==2024.1.0.20240417
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.18
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
